move a function; add tests
